+++ {{cookiecutter.project_name}}/frontend/next.config.js
@@ -2,7 +2,7 @@
 const { i18n } = require('./next-i18next.config');
 
 {% endif -%}
-const { withSentryConfig } = require('@sentry/nextjs');
+const { withSentryConfig } = require('@sentry/nextjs/config');
 
 const basePath = '';